Move autosuggestMaterialSuggestions into a separate helper file
If it lives inside the AutosuggestMaterial.stories.tsx it gets exported
as a separate story, which we don't want.
