+++ src/stories/Library/autosuggest-material/AutosuggestMaterial.stories.tsx
@@ -4,27 +4,7 @@ import {
   AutosuggestMaterial,
   AutosuggestMaterialProps,
 } from "./AutosuggestMaterial";
-
-export const autosuggestMaterialSuggestions = [
-  {
-    cover: "images/book_cover_1.jpg",
-    title: "De uadskillelige",
-    author: "Simone De Beauvoir",
-    year: "1954",
-  },
-  {
-    cover: "images/book_cover_2.jpg",
-    title: "Den lille bog om dansk design",
-    author: "Marie Hugsted",
-    year: "2018",
-  },
-  {
-    cover: "images/book_cover_3.jpg",
-    title: "Audrey Hepburn",
-    author: "Maria Isabel Sanchez Vegara",
-    year: "2018",
-  },
-];
+import { autosuggestMaterialSuggestions } from "./helper";
 
 export default {
   title: "Library / Autosuggest - Material",
