+++ src/stories/Library/autosuggest-material/helper.ts
@@ -1,0 +1,22 @@
+export const autosuggestMaterialSuggestions = [
+  {
+    cover: "images/book_cover_1.jpg",
+    title: "De uadskillelige",
+    author: "Simone De Beauvoir",
+    year: "1954",
+  },
+  {
+    cover: "images/book_cover_2.jpg",
+    title: "Den lille bog om dansk design",
+    author: "Marie Hugsted",
+    year: "2018",
+  },
+  {
+    cover: "images/book_cover_3.jpg",
+    title: "Audrey Hepburn",
+    author: "Maria Isabel Sanchez Vegara",
+    year: "2018",
+  },
+];
+
+export default {};
